Best Subsets Regression

Select the subset of predictors that do the best at meeting some well-defined objective criterion, such as having the largest R2 value or the smallest MSE, Mallow's Cp or AIC.

ols_best_subset(model, ...)

# S3 method for ols_best_subset
plot(x, model = NA, ...)

Arguments

model

an object of class lm

...

other inputs

x

an object of class ols_best_subset

Value

ols_best_subset returns an object of class "ols_best_subset". An object of class "ols_best_subset" is a data frame containing the following components:

References

Kutner, MH, Nachtscheim CJ, Neter J and Li W., 2004, Applied Linear Statistical Models (5th edition). Chicago, IL., McGraw Hill/Irwin.

Examples

## Not run: ------------------------------------
# model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
# ols_best_subset(model)
## ---------------------------------------------

## Not run: ------------------------------------
# # plot
# model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
# k <- ols_best_subset(model)
# plot(k)
## ---------------------------------------------