Hocking's Sp

Average prediction mean squared error.

ols_hsp(model)

Arguments

model
An object of class lm.

Value

Hocking's Sp of the model.

Details

Hocking's Sp criterion is an adjustment of the residual sum of Squares. Minimize this criterion. $$MSE / (n - p - 1)$$

where \(MSE = SSE / (n - p)\), n is the sample size and p is the number of predictors including the intercept

References

Hocking, R. R. (1976). “The Analysis and Selection of Variables in a Linear Regression.” Biometrics 32:1–50.

See also

Other model selection criteria: ols_aic, ols_apc, ols_fpe, ols_mallows_cp, ols_msep, ols_sbc, ols_sbic

Examples

model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
ols_hsp(model)

#> [1] 0.2644378