Ordinary Least Squares Regression

Ordinary Least Squares Regression

ols_regress(object, ...)

# S3 method for lm
ols_regress(object, ...)

Arguments

object

an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted or class lm

...

other inputs

Value

ols_regress returns an object of class "ols_regress". An object of class "ols_regress" is a list containing the following components:

Examples

ols_regress(mpg ~ disp + hp + wt, data = mtcars)

#>                         Model Summary                          
#> --------------------------------------------------------------
#> R                       0.909       RMSE                2.639 
#> R-Squared               0.827       Coef. Var          13.135 
#> Adj. R-Squared          0.808       MSE                 6.964 
#> Pred R-Squared          0.768       MAE                 1.907 
#> --------------------------------------------------------------
#>  RMSE: Root Mean Square Error 
#>  MSE: Mean Square Error 
#>  MAE: Mean Absolute Error 
#> 
#>                              ANOVA                               
#> ----------------------------------------------------------------
#>                 Sum of                                          
#>                Squares    DF    Mean Square      F         Sig. 
#> ----------------------------------------------------------------
#> Regression     931.056     3        310.352    44.565    0.0000 
#> Residual       194.991    28          6.964                     
#> Total         1126.047    31                                    
#> ----------------------------------------------------------------
#> 
#>                                   Parameter Estimates                                    
#> ----------------------------------------------------------------------------------------
#>       model      Beta    Std. Error    Std. Beta      t        Sig      lower     upper 
#> ----------------------------------------------------------------------------------------
#> (Intercept)    37.106         2.111                 17.579    0.000    32.782    41.429 
#>        disp    -0.001         0.010       -0.019    -0.091    0.929    -0.022     0.020 
#>          hp    -0.031         0.011       -0.354    -2.724    0.011    -0.055    -0.008 
#>          wt    -3.801         1.066       -0.617    -3.565    0.001    -5.985    -1.617 
#> ----------------------------------------------------------------------------------------