All possible regression

Fits all regressions involving one regressor, two regressors, three regressors, and so on. It tests all possible subsets of the set of potential independent variables.

ols_step_all_possible(model, ...)

# S3 method for ols_step_all_possible
plot(x, model = NA, ...)

Arguments

model
An object of class lm.

...
Other arguments.

x
An object of class ols_best_subset.

Value

ols_step_all_possible returns an object of class "ols_step_all_possible". An object of class "ols_step_all_possible" is a data frame containing the following components:

n

model number

predictors

predictors in the model

rsquare

rsquare of the model

adjr

adjusted rsquare of the model

predrsq

predicted rsquare of the model

cp

mallow's Cp

aic

akaike information criteria

sbic

sawa bayesian information criteria

sbc

schwarz bayes information criteria

gmsep

estimated MSE of prediction, assuming multivariate normality

jp

final prediction error

pc

amemiya prediction criteria

sp

hocking's Sp

References

Mendenhall William and Sinsich Terry, 2012, A Second Course in Statistics Regression Analysis (7th edition). Prentice Hall

See also

Other variable selection procedures: ols_best_subset, ols_step_backward, ols_step_forward, ols_stepaic_backward, ols_stepaic_both, ols_stepaic_forward

Examples

# NOT RUN {
model <- lm(mpg ~ disp + hp, data = mtcars)
k <- ols_step_all_possible(model)
k
# }

# NOT RUN {
# plot
plot(k)
# }