Stepwise AIC Backward Regression

Build regression model from a set of candidate predictor variables by removing predictors based on Akaike Information Criteria, in a stepwise manner until there is no variable left to remove any more.

ols_stepaic_backward(model, ...)

# S3 method for default
ols_stepaic_backward(model, details = FALSE, ...)

# S3 method for ols_stepaic_backward
plot(x, ...)

Arguments

model

an object of class lm; the model should include all candidate predictor variables

...

other arguments

details

logical; if TRUE, will print the regression result at each step

x

an object of class ols_stepaic_backward

Value

ols_stepaic_backward returns an object of class "ols_stepaic_backward". An object of class "ols_stepaic_backward" is a list containing the following components:

References

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.

Examples

## Not run: ------------------------------------
# # stepwise backward regression
# model <- lm(y ~ ., data = surgical)
# ols_stepaic_backward(model)
## ---------------------------------------------

## Not run: ------------------------------------
# # stepwise backward regression plot
# model <- lm(y ~ ., data = surgical)
# k <- ols_stepaic_backward(model)
# plot(k)
## ---------------------------------------------