Stepwise AIC regression

Build regression model from a set of candidate predictor variables by entering and removing predictors based on akaike information criteria, in a stepwise manner until there is no variable left to enter or remove any more.

ols_stepaic_both(model, details = FALSE)

# S3 method for ols_stepaic_both
plot(x, ...)

Arguments

model
An object of class lm.

details
Logical; if TRUE, details of variable selection will be printed on screen.

x
An object of class ols_stepaic_both.

...
Other arguments.

Value

ols_stepaic_both returns an object of class "ols_stepaic_both". An object of class "ols_stepaic_both" is a list containing the following components:

predictors

variables added/removed from the model

method

addition/deletion

aics

akaike information criteria

ess

error sum of squares

rss

regression sum of squares

rsq

rsquare

arsq

adjusted rsquare

steps

total number of steps

References

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.

See also

Other variable selection procedures: ols_all_subset, ols_best_subset, ols_step_backward, ols_step_forward, ols_stepaic_backward, ols_stepaic_forward

Examples

# NOT RUN {
# stepwise regression
model <- lm(y ~ ., data = stepdata)
ols_stepaic_both(model)
# }

# NOT RUN {
# stepwise regression plot
model <- lm(y ~ ., data = stepdata)
k <- ols_stepaic_both(model)
plot(k)
# }