Stepwise Regression

Build regression model from a set of candidate predictor variables by entering and removing predictors based on p values, in a stepwise manner until there is no variable left to enter or remove any more.

ols_stepwise(model, ...)

# S3 method for default
ols_stepwise(model, pent = 0.1, prem = 0.3,
  details = FALSE, ...)

# S3 method for ols_stepwise
plot(x, model = NA, ...)

Arguments

model
an object of class lm; the model should include all candidate predictor variables

...
other arguments

pent
p value; variables with p value less than pent will enter into the model

prem
p value; variables with p more than prem will be removed from the model

details
logical; if TRUE, will print the regression result at each step

x
an object of class ols_stepwise

Value

ols_stepwise returns an object of class "ols_stepwise". An object of class "ols_stepwise" is a list containing the following components:

orders

candidate predictor variables according to the order by which they were added or removed from the model

method

addition/deletion

steps

total number of steps

predictors

variables retained in the model (after addition)

rsquare

coefficient of determination

aic

akaike information criteria

sbc

bayesian information criteria

sbic

sawa's bayesian information criteria

adjr

adjusted r-square

rmse

root mean square error

mallows_cp

mallow's Cp

indvar

predictors

References

Chatterjee, Samprit and Hadi, Ali. Regression Analysis by Example. 5th ed. N.p.: John Wiley & Sons, 2012. Print.

Examples

# NOT RUN {
# stepwise regression
model <- lm(y ~ ., data = surgical)
ols_stepwise(model)
# }

# NOT RUN {
# stepwise regression plot
model <- lm(y ~ ., data = surgical)
k <- ols_stepwise(model)
plot(k)
# }