Final Prediction Error

Final prediction error

ols_fpe(model)

Arguments

model

an object of class lm

Value

fpe

Details

Computes the estimated mean square error of prediction for each model selected assuming that the values of the regressors are fixed and that the model is correct.

$$MSE((n + p) / n)$$

where \(MSE = SSE / (n - p)\), n is the sample size and p is the number of predictors including the intercept

References

Akaike, H. (1969). “Fitting Autoregressive Models for Prediction.” Annals of the Institute of Statistical Mathematics 21:243–247.

Judge, G. G., Griffiths, W. E., Hill, R. C., and Lee, T.-C. (1980). The Theory and Practice of Econometrics. New York: John Wiley & Sons.

Examples

model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
ols_fpe(model)

#> [1] 7.94966